+++ src/app/app.module.ts
@@ -6,7 +6,6 @@ import { AppComponent } from './app.component';
 import { ProductListModule } from './productListComponent/productListComponent.module';
 import { CartListModule } from './cartListComponent/cartListComponent.module';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { FirstComponentModule } from './firstComponent/firstComponent.module';
 
 @NgModule({
   imports: [
@@ -14,7 +13,6 @@ import { FirstComponentModule } from './firstComponent/firstComponent.module';
     AppRoutingModule,
     ProductListModule,
     CartListModule,
-    FirstComponentModule,
     NgbModule
   ],
   bootstrap: [AppComponent],