+++ src/app/cartListComponent/cartListComponent.module.ts
@@ -1,5 +1,4 @@
 import { NgModule} from '@angular/core';
-import { CommonModule } from '@angular/common'
 
 import { CartListComponent } from './cartListComponent.component';
 import { CartItemModule } from '../cartItemComponent/cartItemComponent.module';
@@ -10,7 +9,6 @@ import { SharedModule } from '../../shared/shared.module';
     CartListComponent
   ],
   imports: [
-    CommonModule,
     CartItemModule,
     SharedModule
   ],