+++ src/core/services/productService.mock.ts
@@ -17,6 +17,25 @@ export const productsListMock: Product[] = [
     id: 3,
     name: 'pensil',
     price: 2,
-    isAvailable: false
+    isAvailable: true
+  },
+  {
+    id: 4,
+    name: 'rule',
+    price: 1,
+    isAvailable: true
+  },
+  {
+    id: 5,
+    name: 'orange',
+    price: 3,
+    isAvailable: true
+  },
+  {
+    id: 6,
+    name: 'apple',
+    price: 2.5,
+    isAvailable: true
   }
+
 ]